chore: rename test file
